feat: Recoil 적용 (#54)
* chore: recoil 설치

* chore: import 문 정리

* feat: recoil root 적용

* refactor: TodoList 컴포넌트에 Recoil 적용

* feat: todo atom

* refactor: Main page와 내부 컴포넌트

* fix: jest config 상대 경로 추가

* chore: home page test 코드에서 렌더링시 RecoilRoot 필요

* chore: Menu 컴포넌트에 임시 타입 지정
