+++ pages/index.tsx
@@ -1,46 +1,17 @@
-import { useEffect, useState, useCallback } from 'react';
 import { NextPage } from 'next';
 import styled from 'styled-components';
 
-import { findAllTodo, deleteTodo } from '@/shared/api/todoAPI';
 import TodoSubmitForm from '@/components/index/TodoSubmitForm';
 import TodoList from '@/components/index/TodoList';
 import SideBar from '@/components/sidebar';
-import { Todo } from '@/shared/types/todo';
 
 const Home: NextPage = () => {
-  const [todoList, setTodoList] = useState<Todo[]>([]);
-
-  useEffect(() => {
-    getAllTodo();
-  }, []);
-
-  const getAllTodo = useCallback(async () => {
-    try {
-      const todos = await findAllTodo();
-      setTodoList(todos);
-    } catch (error) {
-      // !TODO 토스트 팝업
-      alert('투두를 가져오지 못했습니다.');
-    }
-  }, []);
-
-  const onDeleteTodo = useCallback(async (todoId: string) => {
-    try {
-      await deleteTodo(todoId);
-    } catch (error) {
-      alert('todo 삭제 실패');
-    }
-
-    getAllTodo();
-  }, []);
-
   return (
     <Container>
       <SideBar />
       <Content>
-        <TodoList todoList={todoList} onDeleteTodo={onDeleteTodo} />
-        <TodoSubmitForm setTodoList={setTodoList} />
+        <TodoList />
+        <TodoSubmitForm />
       </Content>
     </Container>
   );
