+++ store/todo/atom.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { Todo } from '@/shared/types/todo';
+
+export const todoListState = atom<Todo[]>({
+  key: 'todoListState',
+  default: [],
+});
