+++ tests/pages/index.test.tsx
@@ -1,8 +1,14 @@
 import { render } from '@testing-library/react';
+import { RecoilRoot } from 'recoil';
+
 import Home from 'pages/index';
 
 describe('HomePage', () => {
   test('렌더링', () => {
-    render(<Home />);
+    render(
+      <RecoilRoot>
+        <Home />
+      </RecoilRoot>
+    );
   });
 });
